requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
paddlepaddle
paddleocr==2.7.0.3
pdf2image==1.16.3
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.1.3
python-multipart==0.0.6
requests==2.31.0
Pillow==10.1.0
rapidfuzz==3.5.2
